project/requirements.txt
Flask==0.12.2
google-cloud==0.30.0
google-api-python-client==1.6.4
google-auth-httplib2==0.0.3
PyYAML==3.12
requests==2.18.4
requests_toolbelt==0.8.0
